# api/openapi.yaml
openapi: 3.0.3
info:
  title: PicFast
  version: "1.0"
  description: Image hosting service API
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security: []
paths:
...
  /admin/groups/{id}:
    get:
      tags: [admin]
      summary: Get group detail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
    put:
      tags: [admin]
      summary: Update group
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
      responses:
        "200":
          description: OK
    delete:
      tags: [admin]
      summary: Delete group
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/OK"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    OK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: boolean
        message:
          type: string
        data:
          type: object
...
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
        is_default:
          type: boolean
        is_guest:
          type: boolean
        configs:
          type: object
...